Chapter08/restful_python_2_08_03/Django01/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
coverage==4.5.1
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
httpie==0.9.9
idna==2.7
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.5
py==1.7.0
Pygments==2.2.0
pytest==3.9.3
pytest-cov==2.6.0
pytest-django==3.4.3
pytz==2018.6
requests==2.20.0
six==1.11.0
urllib3==1.24
